Move CameraWaitGroup to camera package
